remove:  region 서비스, member git repo 에서 crew repo 로 이동
