--- src/main/java/hobbiedo/crew/CrewApplication.java
@@ -1,15 +1,0 @@
-package hobbiedo.crew;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class CrewApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CrewApplication.class, args);
-	}
-
-}
