--- src/main/java/hobbiedo/crew/global/config/JpaConfig.java
@@ -1,11 +1,0 @@
-package hobbiedo.crew.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@Configuration
-@EnableJpaAuditing
-@EnableJpaRepositories(basePackages = "hobbiedo.crew.mysqlInfrastructure")
-public class JpaConfig {
-}
