--- src/main/java/hobbiedo/crew/global/exception/GlobalException.java
@@ -1,14 +1,0 @@
-package hobbiedo.crew.global.exception;
-
-import hobbiedo.crew.global.status.ErrorStatus;
-import lombok.Getter;
-
-@Getter
-public class GlobalException extends RuntimeException {
-
-	private final ErrorStatus status;
-
-	public GlobalException(ErrorStatus status) {
-		this.status = status;
-	}
-}
