--- src/main/java/hobbiedo/crew/mongoInfrastructure/ChatRepository.java
@@ -1,20 +1,0 @@
-package hobbiedo.crew.mongoInfrastructure;
-
-import java.time.Instant;
-import java.util.List;
-
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.mongodb.repository.Query;
-
-import hobbiedo.crew.domain.Chat;
-
-public interface ChatRepository extends MongoRepository<Chat, String> {
-	List<Chat> findByCrewId(Long crewId, Pageable pageable);
-
-	@Query("{ 'crewId': ?0, 'imageUrl': { '$exists': true } }")
-	List<Chat> findByCrewIdAndImageUrl(Long crewId);
-
-	@Query(value = "{ 'imageUrl': { '$exists': true }, 'createdAt': { '$lt': ?0 } }", delete = true)
-	void deleteByImageUrlExistsAndCreatedAtBefore(Instant expiryDate);
-}
