+++ src/main/java/hobbiedo/global/config/JpaConfig.java
@@ -1,0 +1,11 @@
+package hobbiedo.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@EnableJpaAuditing
+@EnableJpaRepositories(basePackages = {"hobbiedo.chat.mysqlInfrastructure", "hobbiedo.region.infrastructure"})
+public class JpaConfig {
+}
