+++ src/main/java/hobbiedo/global/config/MongoDbConfig.java
@@ -1,0 +1,33 @@
+package hobbiedo.global.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.config.AbstractMongoClientConfiguration;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.SimpleMongoClientDatabaseFactory;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+
+@Configuration
+@EnableMongoRepositories(basePackages = "hobbiedo.chat.mongoInfrastructure")
+public class MongoDbConfig extends AbstractMongoClientConfiguration {
+
+	@Value("${spring.data.mongodb.uri}")
+	private String mongoUri;
+
+	@Override
+	protected String getDatabaseName() {
+		// URI에 포함된 데이터베이스 이름을 반환
+		return "chat-service";
+	}
+
+	@Bean
+	public MongoTemplate mongoTemplate() throws Exception {
+		return new MongoTemplate(mongoDbFactory());
+	}
+
+	@Bean
+	public SimpleMongoClientDatabaseFactory mongoDbFactory() {
+		return new SimpleMongoClientDatabaseFactory(mongoUri);
+	}
+}
