+++ src/test/java/hobbiedo/chat/CrewApplicationTests.java
@@ -1,0 +1,12 @@
+package hobbiedo.chat;
+
+import org.junit.jupiter.api.Test;
+
+//@SpringBootTest
+class CrewApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
