--- src/test/java/hobbiedo/crew/CrewApplicationTests.java
@@ -1,12 +1,0 @@
-package hobbiedo.crew;
-
-import org.junit.jupiter.api.Test;
-
-//@SpringBootTest
-class CrewApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
